# Conflicts:
#	src/routes/routes.js

+++ src/routes/routes.js
@@ -9,6 +9,7 @@ var mongoClient = mongodb.MongoClient;
 var url = "mongodb://localhost:27017";
 var databaseName = "message_board";
 
+var currentUser = "";
 
 var router = express.Router();
 
@@ -43,12 +44,14 @@ router.route("/").get(
 
                 var posts = await db.collection("messages").find().toArray();
 
+                ;
                 //console.log(posts);
                 var data = {
                     title: "Threads",
                     navOptions: getNav(req.session.user),
                     threads: posts
                 };
+                ;
                 //console.log(data);
                 res.render("index", data);
             } catch (err) {
@@ -67,7 +70,6 @@ router.route("/login").get(
             title: "Log In",
             navOptions: getNav(req.session.user)
         };
-        console.log("USER: " + req.session.user);
         res.render("login", model);
     }
 );
@@ -79,6 +81,7 @@ router.route("/login").post(
                 var client = await mongoClient.connect(url);
                 var db = client.db(databaseName);
                 var user = await db.collection("users").findOne({ "username": req.body.username });
+                currentUser = user;
 
                 if (!user) {
                     res.redirect("/login");
@@ -219,29 +222,31 @@ router.route("/createAccount").post(
 
 //ACCOUNT DETAILS
 router.route("/accountDetails").get(
-    function(req,res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
         console.log(req.session.user);
 
-       (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
+                //var currentUser = await db.collection("users").findOne({/*INSERT QUERY HERE*/});
                 var currentUser = await db.collection("users").findOne({"username": req.session.user.username});
                 
                 var model = {
                     title: "Account Details",
+                    
                     navOptions : getNav(req.session.user),
                     currentUser: currentUser
                 };
                 res.render("accountDetails", model);
-                
-            }catch(err){
+
+            } catch (err) {
                 console.log(err);
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -250,29 +255,30 @@ router.route("/accountDetails").get(
 
 //EDIT ACCOUNT GET
 router.route("/editAccountDetails").get(
-    function(req,res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
         //console.log(req.session.user);
 
-       (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
+                
                 var currentUser = await db.collection("users").findOne({"username":req.session.user.username});
                 
                 var model = {
                     title: "Account Details",
-                    navOptions : getNav(req.session.user),
+                    navOptions: getNav(req.session.user),
                     currentUser: currentUser
                 };
                 res.render("editAccountDetails", model);
-                
-            }catch(err){
+
+            } catch (err) {
                 console.log(err);
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -281,27 +287,28 @@ router.route("/editAccountDetails").get(
 
 //EDIT ACCOUNT POST
 router.route("/editAccountDetails").post(
-    function(req,res){
-        console.log("Changing Account Details!");
+    function (req, res) {
 
-        (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
-                var user = {"_id": "bob" /*USER ID HERE*/};
+                
+                var user = {"username": req.session.user.username};
                 var newValues = {$set: {username: req.body.newUsername, email:req.body.newEmail, age:req.body.newAge}};
                 await db.collection("users").updateOne(user,newValues);
                 req.session.user.username = req.body.newUsername;
                 
 
-            res.redirect("/");
-                
-            }catch(err){
+
+                res.redirect("/");
+
+            } catch (err) {
                 console.log("Mongo Error!");
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -309,11 +316,11 @@ router.route("/editAccountDetails").post(
 )
 
 router.route("/newPost").get(
-    function(req, res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
 
         var model = {
-            title:"Add a new post!",
+            title: "Add a new post!",
             navOptions: getNav(req.session.user)
         };
         res.render("newPost", model);
@@ -321,15 +328,16 @@ router.route("/newPost").get(
 );
 
 router.route("/newPost").post(
-    function(req, res){
+    function (req, res) {
 
-        (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
                 var newMessage = {
+                    
                     "title":req.body.subject,
                     "body":req.body.body,
                     "date_posted":dt.format('m/d/Y'),
@@ -339,10 +347,10 @@ router.route("/newPost").post(
                 await db.collection("messages").insertOne(newMessage);
 
                 res.redirect("/");
-            }catch(err){
+            } catch (err) {
                 console.log("Mongo Error!");
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -351,28 +359,28 @@ router.route("/newPost").post(
 
 //EDIT POST GET
 router.route("/editPost/:name").get(
-    function(req,res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
 
-       (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
-                var currentPost = await db.collection("messages").findOne({"title":req.params.name});
-                
+                var currentPost = await db.collection("messages").findOne({ "title": req.params.name });
+
                 var model = {
                     title: "Account Details",
-                    navOptions : getNav(req.session.user),
+                    navOptions: getNav(req.session.user),
                     currentPost: currentPost
                 };
                 res.render("editPost", model);
-                
-            }catch(err){
+
+            } catch (err) {
                 console.log("Mongo Error!");
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -381,25 +389,24 @@ router.route("/editPost/:name").get(
 
 //POST EDIT POST
 router.route("/editPost/:name").post(
-    function(req,res){
-        console.log("Updating Post");
+    function (req, res) {
 
-       (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
-                var post = {"name": req.params.name};
-                var newValues = {$set: {name: req.body.editName, body: req.body.editThread}};
-                await db.collection("users").updateOne(post,newValues);
+                var post = { "name": req.params.name };
+                var newValues = { $set: { name: req.body.editName, body: req.body.editThread } };
+                await db.collection("users").updateOne(post, newValues);
 
-                
-                res.redirect("/");   
-            }catch(err){
+
+                res.redirect("/");
+            } catch (err) {
                 console.log("Mongo Error!");
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -411,19 +418,27 @@ router.route("/deletePost/:title").get(
         console.log("Delete Post!");
         //console.log(req.params);
 
-        (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
-                await db.collection("messages").deleteOne({"title":req.params.name});
+                if (currentUser.type.includes("admin")) {
+                    await db.collection("messages").deleteOne({ "title": req.params.name });
+                } else {
+                    var post = await db.collection("messages").findOne({ "title": req.params.name });
+                    if (post.createdBy == currentUser.username) {
+                        await db.collection("messages").deleteOne({ "title": req.params.name });
+                    }
+                }
+
 
                 res.redirect("/");
-            }catch(err){
+            } catch (err) {
                 console.log("Mongo Error!");
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -431,30 +446,29 @@ router.route("/deletePost/:title").get(
 );
 
 router.route("/post/:name").get(
-    function(req, res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
-        (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
 
-                var post = await db.collection("messages").findOne({"title":req.params.name});
-                var userInfo = await db.collection("users").findOne({"username":post.createdBy});
+                var post = await db.collection("messages").findOne({ "title": req.params.name });
+                var userInfo = await db.collection("users").findOne({ "username": post.createdBy });
 
-                // console.log(userInfo);
                 var model = {
                     title: "Post Detail Page",
-                    navOptions : getNav(req.session.user),
+                    navOptions: getNav(req.session.user),
                     message: post,
                     user: userInfo
                 };
                 res.render("viewPost", model);
-            }catch(err){
+            } catch (err) {
                 console.log("Mongo Error!");
                 console.log(err);
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
@@ -462,10 +476,10 @@ router.route("/post/:name").get(
 );
 
 router.route("/posts").get(
-    function(req, res){
+    function (req, res) {
         var getNav = req.app.get("getNav");
-        (async function mongo(){
-            try{
+        (async function mongo() {
+            try {
                 var client = await mongoClient.connect(url);
 
                 var db = client.db(databaseName);
@@ -474,16 +488,18 @@ router.route("/posts").get(
 
                 var model = {
                     title: "Post List",
-                    navOptions : getNav(req.session.user),
+                    navOptions: getNav(req.session.user),
                     posts: posts
                 };
                 res.render("posts", model);
-            }catch(err){
+            } catch (err) {
                 res.send(err);
-            }finally{
+            } finally {
                 client.close();
             }
         }());
     }
 );
+
+
 module.exports = router;
